Created the Coding
Completed the Assignment- The code imports urllib, json and sll. It prompts the user for a url and if the user does not add a url (<1 character in length), it uses an included url. Next, it creates the connection and reads the JSON data from the URL using urllib. It parses and extracts the comment counts from the JSON data, loops through the comments, and computes the sum of the numbers in the ‘comments’ in the file. Finally, it prints the resulting count and sum.